+++ sovtimer/__init__.py
@@ -2,5 +2,5 @@
 a couple of variable to use throughout the app
 """
 
-__version__ = "1.5.0"
+__version__ = "1.5.1"
 __title__ = "Sovereignty Timers"