+++ CHANGELOG.md
@@ -9,6 +9,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [In Development] - Unreleased
 
 
+## [1.5.1] - 2022-05-25
+
+### Added
+
+- Versioned static files to break the browser cache on app update
+
+### Changed
+
+- JavaScript modernized
+
+### Fixed
+
+- Attackers default score in DB
+
+
 ## [1.5.0] - 2022-03-02
 
 ### Added